Addressed comments of @reenl.
The maximum sizes of a value and a split value are now defined, and the
checking logic for split values and the number of split parts is
extracted into their own methods.

Also, these methods now check for the integrity of values, which in the
case of ::getNumberOfParts() prevents a possible DoS on the Memcached
server.

Signed-off-by: Jacob Kiers <jacob@alphacomm.nl>