Improve scalability of Worker Handler
The Worker Handler dispatched request by looping through the
requests. When there are many requests, this causes a timeout
on workers for which there are no (or only a few) requests.

Therefore, the Worker Handler will now dispatch the requests
by looping through the requested actions. Since this list
is usually significantly shorter, the timeouts go away.