Prevent looping every request on ::queueStatus()
When requesting the queue status, the ::queueStatus() function looped
through every request to determine the number of requests for each
action.

However, with the last fix in the Worker Handler, all requests are
already grouped by the action name.

Therefore, it is now just a simple matter of counting the requests.

Signed-off-by: Jacob Kiers <jacob@alphacomm.nl>