Allow a fetch without an execute
A request is fetched based on its request ID. In order to allow truly
asynchronous fetches, it is necessary to make sure that a fetch can be
performed when the Manager is not yet known.

This situation happens when a new Request object is created just for
fetching an earlier request, without knowing the service, arguments or
even cache key of the request.

Signed-off-by: Jacob Kiers <jacob@alphacomm.nl>